[TP1] Corrigindo os valores do limiar partição

+++ TP1/include/ordenador.hpp
@@ -7,7 +7,7 @@ using namespace std;
 
 class OrdenadorUniversal{
   private:
-    float a, b, c;
+    double a, b, c;
     int numQuebras;
     int seed;
     int *vetorOrig;
@@ -19,17 +19,17 @@ class OrdenadorUniversal{
     int _passoMPS(int max, int min);
     int menorCustoMinParticao(int numMPS);
     int menorCustoLimiarQuebras(int numMPS, SortingAlgorithm *sort);
-    float calcularCusto(int cmp, int move, int calls);
+    double calcularCusto(int cmp, int move, int calls);
     void ordenadorUniversal(int vetor[], int tam, int minTamParticao, int limiarQuebras);
     void registraEstatisticasMinParticao(int tamMin, int numMPS);
     void registraEstatisticasLimQuebras(int breaks, int numMPS, SortingAlgorithm algo);
-    void calculaNovaFaixa(float limParticao, int* minMPS, int* maxMPS, int* passoMPS, int* numMPS, int *minNumMPS, int *maxNumMPS);
+    void calculaNovaFaixa(double limParticao, int* minMPS, int* maxMPS, int* passoMPS, int* numMPS, int *minNumMPS, int *maxNumMPS);
     void imprimeEstatisticasMinParticao(Registro r);
     void imprimeEstatisticasLimiarQuebras(int numMPS, SortingAlgorithm algo);
     void copiaVetor(int vetor[], int tam);
 
   public:
-    OrdenadorUniversal(float a, float b, float c, int numQuebras, int seed, int* vetorOrig)
+    OrdenadorUniversal(double a, double b, double c, int numQuebras, int seed, int* vetorOrig)
     : a(a), b(b), c(c), numQuebras(numQuebras), seed(seed), vetorOrig(vetorOrig) { }
     OrdenadorUniversal();
     int determinarLimiarParticao(int vetor[], int tam, int limiarCusto);
